Use dbus-send to query for remaining time for non-OSX
Unfortunately `upower` does not return the remaining time in seconds and
instead uses formatted strings like "42 minutes". This makes it hard to
add the time of many batteries together. Fortunately UPower also
provides a dbus device that can be queried using dbus-send.

I had to write some helper functions to make this work. This increased
the complexity compared to the previous solution. That's why I decided
to also write a minimal test framework to ensure that they can be
refactored safely.